Create README.md
Hawkes Process (HP) and Non-Homogeneous Poisson Process (NHPP) simulation Algorithms 
There are 6 R codes, 3 of them are Hawkes Process Simulation, 3 of them are for Non-Homogeneous Poisson Process Simulation
The Hawkes Process is simulated with thinning algorithm and with composition algorithm
For the composition algorithm, the inversion method with transformation method and the inversion method with transformation method combined with order statistics
are used. The numerical inversion is used to invert cumulative intensity function. The Spline interpolation and UNU.RAN Random Variate Generators are used.
The Non-Homogeneous Poisson Process (NHPP) is simulated with thinning algorithm and the inversion method with transformation method and 
the inversion method with transformation method combined with order statistics.